+++ pkg/conf/auth.go
@@ -1,9 +1,12 @@
 package conf
 
-import "github.com/sourcegraph/sourcegraph/schema"
+import (
+	"github.com/sourcegraph/sourcegraph/schema"
+)
 
 // AuthProviderType returns the type string for the auth provider.
-func AuthProviderType(p schema.AuthProviders) string {
+func AuthProviderType() string {
+	p := AuthProvider()
 	switch {
 	case p.Builtin != nil:
 		return p.Builtin.Type